# Randomized allocation sampling