--- a/src/coreclr/nativeaot/Runtime/thread.inl
+++ b/src/coreclr/nativeaot/Runtime/thread.inl
@@ -3,7 +3,9 @@
 
 #ifndef DACCESS_COMPILE
 
+#include "eventtracebase.h"
 
+const uint32_t SamplingDistributionMean = (100 * 1024);
 
 inline gc_alloc_context* ee_alloc_context::GetGCAllocContext()
 {
@@ -22,11 +24,49 @@ struct _thread_inl_gc_alloc_context
     uint8_t* alloc_limit;
 };
 
-inline void ee_alloc_context::UpdateCombinedLimit()
+
+inline bool ee_alloc_context::IsRandomizedSamplingEnabled()
 {
-    // The randomized allocation sampling feature is being submitted in stages. For now sampling is never enabled so
-    // combined_limit is always the same as alloc_limit.
-    combined_limit = ((_thread_inl_gc_alloc_context*)GetGCAllocContext())->alloc_limit;
+#ifdef FEATURE_EVENT_TRACE
+    return IsRuntimeProviderEnabled(TRACE_LEVEL_INFORMATION, CLR_ALLOCATIONSAMPLING_KEYWORD);
+#else
+    return false;
+#endif // FEATURE_EVENT_TRACE
+}
+
+inline void ee_alloc_context::UpdateCombinedLimit(bool samplingEnabled)
+{
+    _thread_inl_gc_alloc_context* gc_alloc_context = (_thread_inl_gc_alloc_context*)GetGCAllocContext();
+    if (!samplingEnabled)
+    {
+        combined_limit = gc_alloc_context->alloc_limit;
+    }
+    else
+    {
+        // compute the next sampling budget based on a geometric distribution
+        size_t samplingBudget = ComputeGeometricRandom();
+
+        // if the sampling limit is larger than the allocation context, no sampling will occur in this AC
+        // We do Min() prior to adding to alloc_ptr to ensure alloc_ptr+samplingBudget doesn't cause an overflow.
+        
+        size_t size = gc_alloc_context->alloc_limit - gc_alloc_context->alloc_ptr;
+        combined_limit = gc_alloc_context->alloc_ptr + min(samplingBudget, size);
+    }
+}
+
+inline uint32_t ee_alloc_context::ComputeGeometricRandom()
+{
+    // compute a random sample from the Geometric distribution.
+    double probability = t_random.NextDouble();
+    uint32_t threshold = (uint32_t)(-log(1 - probability) * SamplingDistributionMean);
+    return threshold;
+}
+
+// Returns a random double in the range [0, 1).
+inline double ee_alloc_context::PerThreadRandom::NextDouble()
+{
+    uint32_t value = minipal_xoshiro128pp_next(&random_state);
+    return value * (1.0/(UINT32_MAX+1.0));
 }
 
 // Set the m_pDeferredTransitionFrame field for GC allocation helpers that setup transition frame