… docs/design/features/RandomizedAllocationSampling.md
… src/coreclr/inc/eventtracebase.h
… src/coreclr/minipal/Unix/CMakeLists.txt
… src/coreclr/minipal/Windows/CMakeLists.txt
… src/coreclr/nativeaot/Runtime/CMakeLists.txt
… src/native/minipal/CMakeLists.txt
… src/native/minipal/xoshiro128pp.c
… src/native/minipal/xoshiro128pp.h
… src/tests/tracing/eventpipe/randomizedallocationsampling/allocationsampling.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/allocationsampling.csproj
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/Allocate.csproj_
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/AllocateArraysOfDoubles.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/AllocateDifferentTypes.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/AllocateRatioSizedArrays.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/AllocateSmallAndBig.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/AllocationsRunEventSource.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/IAllocations.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/Program.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/Allocate/ThreadedAllocations.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/AllocationProfiler.sln
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/AllocationProfiler/AllocationProfiler.csproj_
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/AllocationProfiler/Program.cs
… src/tests/tracing/eventpipe/randomizedallocationsampling/manual/README.md