@noahfalk 2024-07-20
Functional testing found and fixed an off-by-one error in the RNG code but otherwise things looked fine. I also resynced this PR on top of the latest changes in https://github.com/dotnet/runtime/pull/104849 and https://github.com/dotnet/runtime/pull/104851. The last commit, now number 10, remains the interesting one.

I also did some performance testing using GCPerfSim as an allocation benchmark. My default configuration was 4 threads, workstation mode, 500GB of allocations entirely with small objects and no survival. It is intended to put maximum stress on the allocation code paths. GCPerfSim command line: `-tc 4 -tagb 500 -tlgb 0.05 -lohar 0 -lohsr 100000-2000000 -sohsi 0 -lohsi 0 -pohsi 0 -sohpi 0 -lohpi 0 -sohfi 0 -lohfi 0 -pohfi 0 -allocType reference -testKind time`. I also ran a few variations that added modest amounts of survival and LOH allocations that are a bit more realistic, though still extremely allocation heavy.

### Benchmarks - No Tracing enabled

| Scenario                                    |    Baseline time   |    PR  time   |
|---------------------------------------------|--------------------|---------------|
| Default                                     |     20.8           |    21.4       |
| Default + `sohsi 100`                       |     40.9           |    41.8       |
| Default + `sohsi 100 -lohar 10`             |     42.8           |    43.9       |
| Default + `sohsi 100 -lohar 10 -lohsi 200`  |     43.6           |    44.6       |

### Benchmarks - Tracing AllocSampling+GC keywords, verbose level

| Scenario                                    |    Baseline time   |    PR  time   |
|---------------------------------------------|--------------------|---------------|
| Default                                     |     22             |    23.1       |
| Default + `sohsi 100`                       |     41.6           |    43.0       |
| Default + `sohsi 100 -lohar 10`             |     44.7           |    44.5       |
| Default + `sohsi 100 -lohar 10 -lohsi 200`  |     44.2           |    45.0       |

Overall it looks like around ~0.9 additional seconds for the PR to do 500GB of allocations. On a tight microbenchmark its noticeable and then as other GC or non-allocation costs increase it becomes relatively less noticeable. I'm investigating to see if it can be improved at all.