@jkotas 2024-07-29
- GC stress run identified a functional issue in https://github.com/dotnet/runtime/pull/104849: https://github.com/dotnet/runtime/pull/104849#issuecomment-2254934408

- https://github.com/dotnet/runtime/pull/104851 is still running native AOT tests. Also, there is a question for Mike about the data contract.