@noahfalk 2024-07-29
> Given that we are feature complete for .NET 9, should this get an approval from Jeff or tactics before merging for .NET 9?

Yes, I was assuming that would be part of the process.