@noahfalk 2024-10-31
> I'll run my tests on monday - the rest of the week is off for me sorry

Sounds great! Its not a rush :)