@jkotas 2024-11-07
I do not have any additional feedback. As I have mentioned in https://github.com/dotnet/runtime/pull/104955#pullrequestreview-2205426994, this should get a review from somebody familiar with tracing.