@noahfalk 2024-11-07
> As I have mentioned in https://github.com/dotnet/runtime/pull/104955#pullrequestreview-2205426994, this should get a review from somebody familiar with tracing.

@brianrob - would you mind taking a look if you haven't already?